Add files via upload
The scope of this Assignment is to get acquainted with JavaScript programming and Google Maps API. Tasks:
Create an application using Google Maps API that will represent on top of Google Maps the content of the KML file you created have already. Attention: this is not a visualizer of the KML file. You must encode the same content in an HTML/JS file using simple Google Maps API geometry types, such as Markers, Polylines, Polygons, etc.