Create A1
The scope of this Assignment is to get acquainted with the basic HTML encoding. Complete the following two tasks:

Using a text editor, compile an HTML page (e.g., your personal page or a project home page). Include in this page: hyperlinks, images, lists, tables; and change the default formatting parameters using appropriate styles.
Include in the same page an image or map with clickable areas and hyperlinks.