Create SVG & KML
The scope of this Assignment is to get acquainted with SVG and KML languages. Complete the following two tasks:

Enrich the image map you used in Assignment 1 with vector items encoded in SVG: Put the image as background and draw on top some simple elements (lines, polygons, etc.). Create hyperlinks on these items and a map legend. Draw a couple of suggested visiting paths and simulate the visit with animated objects.
Encode the above map in KML on top of Google Maps. Draft the KML file manually in a text edtor. Extract the coordinate values (lat,long) of various locations (manually) from Google Earth. In the Settings, you can request the coordinates of the cursor to be reported in decimal degrees (in the lower right corner of the screen). No need to collect too many locations. You must include three or more points of interest (<Placemark>, <Point>) and a rough version of paths and/or polygons connecting/enclosing them (<Placemark>, <LinearRing> or <Polygon>). Attach relevant information, such as text descriptions, images and hyperlinks to the placemarks. Do not encode any animation or 3D structure in the KML file.