Create A4
Using php create the following application. The user will be prompted to type a latitude value within the interval [-90,+90] in a form field. After submitting the value to the server, the length (circumference) of the parallel in that latitude will be returned. The application will provide an error message if the latitude value is not valid. The format of latitude values is signed decimal degrees (+dd.ddd). Make the assumption of a spherical earth with Radius=6,371km.Circumference of parallel 
Create a form where the user can type three pairs of coordinates (latitude and longitude) for three locations on Earth. These coordinates will then be used by a php script to generate a new page to draw the spherical triangle defined by these locations on top of Google Maps. See https://developers.google.com/maps/documentation/javascript/examples/polyline-simple Use this example code to draw the three edges of the triangle.